Following new convention and putting childUplift intervention in the child_poverty_interventions.py script
